add Output for controlling the GP8XXX 2-channel DAC (0-10 VDC)
